moved CommentList to components
